requirements.txt
anyio==3.7.1
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2025.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
tornado==6.2
typing_extensions==4.7.1
tzdata==2025.1
tzlocal==5.1
urllib3==1.26.6
